Squelch warning about an integer overflow
We cannot rely on long integers to have more than 32 bits.


cregit: 8f7744235844c44e780e026868beef29693c062e http://github.com/git/git/commit/8f7744235844c44e780e026868beef29693c062e

+++ git-compat-util.h
@@ -2382,7 +2382,7 @@ define|#
 directive|define
 name|DEFAULT_PACKED_GIT_LIMIT
 define|\
-value|((1024L * 1024L) * (sizeof(void*)>= 8 ? 8192 : 256))
+value|((1024L * 1024L) * (size_t)(sizeof(void*)>= 8 ? 8192 : 256))
 end_define
 begin_ifdef
 ifdef|#
